Update build script

+++ build.sh
@@ -4,7 +4,7 @@ set -eux
 BUILD_USER_HOME="${BUILD_USER_HOME:-/build}"
 BUILD_USER_NAME="${BUILD_USER_NAME:-build}"
 # Debian release used during build
-DEBIAN_RELEASE="${DEBIAN_RELEASE:-stretch}"
+DEBIAN_RELEASE="${DEBIAN_RELEASE:-bookworm}"
 # Mattermost version to build
 MATTERMOST_RELEASE="${MATTERMOST_RELEASE:-v5.26.0}"
 MMCTL_RELEASE="${MMCTL_RELEASE:-v5.26.0}"
@@ -47,16 +47,23 @@ if [ "$(id -u)" -eq 0 ]; then # as root user
 	apt-get build-dep --quiet \
 		pngquant
 	# install go from golang.org
-	wget https://golang.org/dl/go${GO_VERSION}.linux-amd64.tar.gz
-	tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
-	mv go /usr/local
+	if [ ! -f go${GO_VERSION}.linux-amd64.tar.gz ]; then
+		wget https://golang.org/dl/go${GO_VERSION}.linux-amd64.tar.gz
+	fi
+	if [ ! -d /usr/local/go ]; then
+		tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
+		mv go /usr/local
+	fi
 	export GOROOT=/usr/local/go
 	export PATH=$GOROOT/bin:$PATH
 	# FIXME go (executed by build user) writes to GOROOT
 	install --directory --owner="${BUILD_USER_NAME}" \
 		"$(go env GOROOT)/pkg/$(go env GOOS)_$(go env GOARCH)"
+
+	echo "GOOS_GOARCH=$(go env GOOS)_$(go env GOARCH)"
 	# Re-invoke this build.sh script with the 'build' user
 	runuser -u "${BUILD_USER_NAME}" -- "${0}"
+
 	# salvage build artifacts
 	cp --verbose \
 		"${BUILD_USER_HOME}/mattermost-${MATTERMOST_RELEASE}-$(go env GOOS)-$(go env GOARCH).tar.gz" \
@@ -95,31 +102,18 @@ popd
 
 # prepare the go build environment
 install --directory "${HOME}/go/bin"
-if [ "$(go env GOOS)_$(go env GOARCH)" != 'linux_amd64' && ! -f ${HOME}/go/bin/linux_amd64 ]; then
+if [ "$(go env GOOS)_$(go env GOARCH)" != 'linux_amd64' ] && [ ! -f "${HOME}/go/bin/linux_amd64" ] && [ ! -h "${HOME}/go/bin/linux_amd64" ]; then
 	ln --symbolic \
 		"${HOME}/go/bin/$(go env GOOS)_$(go env GOARCH)" \
 		"${HOME}/go/bin/linux_amd64"
 fi
-# build mmctl
-install --directory "${HOME}/go/src/github.com/mattermost/mmctl"
-wget --quiet --continue --output-document="mmctl.tar.gz" \
-	"https://github.com/mattermost/mmctl/archive/${MMCTL_RELEASE}.tar.gz"
-tar --directory="${HOME}/go/src/github.com/mattermost/mmctl" \
-	--strip-components=1 --extract --file="mmctl.tar.gz"
-find "${HOME}/go/src/github.com/mattermost/mmctl/" -type f -name '*.go' | xargs \
-	sed -i \
-	-e 's#//go:build linux || darwin#//go:build linux || darwin || dragonfly || freebsd || netbsd || openbsd#' \
-	-e 's#// +build linux darwin#// +build linux darwin dragonfly freebsd netbsd openbsd#'
-make --directory="${HOME}/go/src/github.com/mattermost/mmctl" \
-	BUILD_NUMBER="dev-$(go env GOOS)-$(go env GOARCH)-${MMCTL_RELEASE}" \
-	ADVANCED_VET=0 \
-	GO="GOARCH= GOOS= $(command -v go)"
+
 # build focalboard
 make --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
 	prebuild
 make --directory="${HOME}/go/src/github.com/mattermost/focalboard" \
 	build
-# build Mattermost webapp
+# # build Mattermost webapp
 npm set progress false
 sed -i -e 's#--verbose#--display minimal#' \
 	"${HOME}/go/src/github.com/mattermost/mattermost/webapp/package.json"
@@ -127,7 +121,7 @@ make --directory="${HOME}/go/src/github.com/mattermost/mattermost/webapp" \
 	dist
 # build Mattermost server
 patch --directory="${HOME}/go/src/github.com/mattermost/mattermost/server" \
-	--strip=1 -t < "${HOME}/build-release.patch"
+	--strip=1 -t < "/root/build-release.patch"
 sed -i \
 	-e 's#go generate#env --unset=GOOS --unset=GOARCH go generate#' \
 	-e 's#$(GO) generate#env --unset=GOOS --unset=GOARCH go generate#' \
